refactor(fast-usdc): FastUSDCConfigShape -> meta.deployConfigShape
